refactor gene components
